Black And Blue Color Letters
There was a issue in the last program to print in PDF format, It stuck to the loading screen, because of Filter property used in CSS thats making it slow to process each letter one by one, So here i fixed it with different color letters BLACK AND BLUE separately, which eliminate the Filter property used. Now you can print in PDF too.